requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pyairtable
notion-client
